backend/requirements.txt
Django
djangorestframework
mysqlclient
django-mysql
django-cors-headers
